% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{con_check}
\alias{con_check}
\title{Check connectedness of multiple factors in a dataframe}
\usage{
con_check(data, formula, WW = TRUE)
}
\arguments{
\item{data}{A dataframe}

\item{formula}{A formula with multiple factor names in the dataframe.}

\item{WW}{Pass-through argument to \code{compfactor}.}
}
\value{
A vector with integers representing the group membership of each observation.
}
\description{
Multiple factors in a dataframe are said to be connected if a model matrix
based on those factors is full rank.

This function provides a formula interface to the lfe::compfactor function
to check for connectedness.
}
\examples{
\dontrun{
# In the data_eccleston dataframe, each pair of factors is connected.
con_check(data_eccleston, ~ row + trt)
con_check(data_eccleston, ~ col + trt)
con_check(data_eccleston, ~ row + col)
# But all three factors are COMPLETELY disconnected into 16 groups.
con_check(data_eccleston, ~ row + col + trt)
}

}
\references{
None
}
\author{
Kevin Wright
}
