% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{con_view}
\alias{con_view}
\title{View connectedness of two factors in a dataframe with a levelplot}
\usage{
con_view(
  data,
  formula,
  fun.aggregate = mean,
  na.rm = TRUE,
  xlab = "",
  ylab = "",
  cex.num = 0.75,
  cex.x = 0.7,
  cex.y = 0.7,
  col.regions = RedGrayBlue,
  cluster = "incidence",
  dropNA = TRUE,
  ...
)
}
\arguments{
\item{data}{Name of a data.frame}

\item{formula}{A formula like \code{yield ~ fx*fy}}

\item{fun.aggregate}{The function to use for aggregating data in cells. Default is mean.}

\item{na.rm}{Should NA values be removed? Default is TRUE.}

\item{xlab}{Label for x axis}

\item{ylab}{Label for y axis}

\item{cex.num}{Disjoint group number.}

\item{cex.x}{Scale factor for x axis tick labels.  Default 0.7.}

\item{cex.y}{Scale factor for y axis tick labels  Default 0.7.}

\item{col.regions}{Function for color regions. Default RedGrayBlue.}

\item{cluster}{If "incidence", cluster rows and columns by the incidence matrix.}

\item{dropNA}{If TRUE, observed data that are \code{NA} will be dropped.}

\item{...}{Other parameters.}
}
\description{
If there is replication for the treatment combination cells in a2
two-way table, the replications are averaged together (or counted)
before constructing the heatmap.

By default, rows and columns are clustered using the 'incidence' matrix of 0s and 1s.

The function checks to see if the cells in the heatmap form a connected
set.  If not, the data is divided into connected subsets and the subset
group number is shown within each cell.

By default, missing values in the response are deleted.
}
\examples{

require(lattice)
bar = transform(lattice::barley, env=factor(paste(site,year)))
set.seed(123)
bar <- bar[sample(1:nrow(bar), 70, replace=TRUE),]
con_view(bar, yield ~ variety * env, cex.x=1, cex.y=.3, cluster=FALSE)

# Create a heatmap of cell counts
w2b = colorRampPalette(c('wheat','black'))
con_view(bar, yield ~ variety * env, fun.aggregate=length,
  cex.x=1, cex.y=.3, col.regions=w2b, cluster=FALSE)

# Example from paper by Fernando et al. (1983).
set.seed(42)
data_fernando = transform(data_fernando,
  y=stats::rnorm(9, mean=100))
con_view(data_fernando, y ~ gen*herd, cluster=FALSE,
     main = "Fernando unsorted")
con_view(data_fernando, y ~ gen*herd, cluster=TRUE,
     main = "Fernando unsorted")

# Example from Searle (1971), Linear Models, p. 325
dat2 = transform(data_searle,
  y=stats::rnorm(nrow(data_searle)) + 100)

con_view(dat2, y ~ f1*f2, cluster=FALSE, main="data_searle unsorted") 
con_view(dat2, y ~ f1*f2, main="data_searle clustered")

}
\author{
Kevin Wright
}
