% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{data_searle}
\alias{data_searle}
\title{Data from Searle}
\format{
An object of class \code{data.frame} with 14 rows and 2 columns.
}
\source{
Searle (1971).
Linear Models. Page 324.
}
\usage{
data_searle
}
\description{
Data from Searle
}
\details{
A dataframe with 2 treatment factors.
The treatment combinations form 3 disconnected groups.
}
\examples{
cbind(data_searle,
      .group=con_check(data_searle, ~ f1 + f2))
data_searle = transform(data_searle,
  y = rnorm(nrow(data_searle), mean=100))
con_view(data_searle, y ~ f1*f2, cluster=FALSE, main="Searle unsorted")
con_view(data_searle, y ~ f1*f2, main="Searle clustered")

}
\keyword{datasets}
