% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{data_weeks1}
\alias{data_weeks1}
\title{Data from Weeks & Williams example 1}
\format{
An object of class \code{data.frame} with 16 rows and 3 columns.
}
\source{
Weeks, David L. & Donald R. Williams (1964).
A Note on the Determination of Connectedness in an N-Way Cross Classification.
Technometrics, 6:3, 319-324. Table 1.
http://dx.doi.org/10.1080/00401706.1964.10490188
}
\usage{
data_weeks1
}
\description{
Data from Weeks & Williams example 1
}
\details{
A dataframe with 3 treatment factors.
The treatment combinations are connected.

Note: This data is based on Table 1 of Weeks & Williams.
Table 2 is missing treatment combination (1,2,4).
}
\examples{
library(lfe)
cbind(data_weeks1,
      .group=con_check(data_weeks1, ~ f1+f2+f3))

}
\keyword{datasets}
