% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{data_weeks2}
\alias{data_weeks2}
\title{Data from Weeks & Williams example 2}
\format{
An object of class \code{data.frame} with 62 rows and 3 columns.
}
\source{
Weeks, David L. & Donald R. Williams (1964).
A Note on the Determination of Connectedness in an N-Way Cross Classification.
Technometrics, 6:3, 319-324. Table 3.
http://dx.doi.org/10.1080/00401706.1964.10490188
}
\usage{
data_weeks2
}
\description{
Data from Weeks & Williams example 2
}
\details{
A dataframe with 3 treatment factors.
The treatment combinations form 4 disconnected groups.

Note: This data is based on Table 3 of Weeks & Williams.
The groups defined in the text are missing some combinations.
}
\examples{
library(lfe)
cbind(data_weeks2,
      .group=con_check(data_weeks2, ~f1 + f2 + f3))

}
\keyword{datasets}
